timer: more safety

+++ src/timer.h
@@ -10,8 +10,8 @@ void clock_init_fast();
 
 void timer_init();
 
-void timer_timeout_set_100us(uint32_t us100);
-void timer_timeout_set_ms(uint32_t ms);
+void timer_timeout_set_100us(uint16_t us100);
+void timer_timeout_set_ms(uint16_t ms);
 
 uint8_t timer_timeout();
